pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/color_calibration"]

[project]
name = "color-calibration"
dynamic = ["version"]
description = "Tools for calibrating color in images"
authors = [
    { name = "David Carlyn", email = "davidecarlyn@gmail.com" },
    { name = "Matthew J. Thompson", email = "thompson.m.j@outlook.com"}
]
readme = "README.md"
license = { text = "MIT" }
keywords = []
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "numpy",
    "opencv-python",
    "pycocotools",
    "tqdm",
    "plotly",
    "Pillow"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
]

[project.urls]
"Source" = "https://github.com/Imageomics/color-calibration"
"Issues" = "https://github.com/Imageomics/color-calibration/issues"

[project.scripts]
colorcal = "color_calibration.__main__:main"

[tool.hatch.build]
include = [
    "README.md",
    "LICENSE"
]

[tool.hatch.version]
path = "src/color_calibration/__init__.py"
